parser/requirements.txt
Pillow
Flask
webdriver_manager
seleniumbase
python-dotenv